#335: Made changes according to review
-
