+++ src/app/ffbe/ffbe.module.ts
@@ -52,8 +52,8 @@ import {LatentSkillsComponent} from './latent-skills/latent-skills.component';
 import {LatentSkillsService} from './services/latent-skills.service';
 import {CaracteristiquesContainerDisplayComponent} from './caracteristiques-container-display/caracteristiques-container-display.component';
 import {CaracteristiquesEXDisplayComponent} from './caracteristiques-ex-display/caracteristiques-ex-display.component';
-import { ObjetTueursDuoDisplayComponent } from './objets-display-components/objet-tueurs-duo-display/objet-tueurs-duo-display.component';
-import { ObjetTueursDisplayComponent } from './objets-display-components/objet-tueurs-display/objet-tueurs-display.component';
+import {ObjetTueursPhysiquesMagiquesDisplayComponent} from './objets-display-components/objet-tueurs-physiques-magiques-display/objet-tueurs-physiques-magiques-display.component';
+import {ObjetTueursDisplayComponent} from './objets-display-components/objet-tueurs-display/objet-tueurs-display.component';
 
 @NgModule({
   imports: [
@@ -101,7 +101,7 @@ import { ObjetTueursDisplayComponent } from './objets-display-components/objet-t
     LatentSkillsComponent,
     CaracteristiquesContainerDisplayComponent,
     CaracteristiquesEXDisplayComponent,
-    ObjetTueursDuoDisplayComponent,
+    ObjetTueursPhysiquesMagiquesDisplayComponent,
     ObjetTueursDisplayComponent,
   ],
   exports: [
