+++ src/app/ffbe/objets-display-components/objet-display/objet-display.component.html
@@ -73,7 +73,7 @@
 
     <app-objet-alterations-etat-duo-display [objet]="objet"></app-objet-alterations-etat-duo-display>
 
-    <app-objet-tueurs-duo-display [objet]="objet"></app-objet-tueurs-duo-display>
+    <app-objet-tueurs-physiques-magiques-display [objet]="objet"></app-objet-tueurs-physiques-magiques-display>
 
     <app-skills-display *ngIf="areObjetCompetencesDisplayed()"
                         [competences]="objet.competences"
