+++ src/app/ffbe/objets-display-components/objet-tueurs-display/objet-tueurs-display.component.css
@@ -1,0 +1,3 @@
+th {
+  text-transform: capitalize;
+}
