+++ src/app/ffbe/objets-display-components/objet-tueurs-display/objet-tueurs-display.component.html
@@ -7,51 +7,51 @@
     <table>
       <caption>Tueurs</caption>
       <tr>
-        <th>{{formateMonsterType(2)}}</th>
+        <th>{{getMonsterName(2)}}</th>
         <td>{{formateTueur(tueurs.aquatiques)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(0)}}</th>
+        <th>{{getMonsterName(0)}}</th>
         <td>{{formateTueur(tueurs.betes)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(1)}}</th>
+        <th>{{getMonsterName(1)}}</th>
         <td>{{formateTueur(tueurs.oiseaux)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(3)}}</th>
+        <th>{{getMonsterName(3)}}</th>
         <td>{{formateTueur(tueurs.demons)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(6)}}</th>
+        <th>{{getMonsterName(6)}}</th>
         <td>{{formateTueur(tueurs.dragons)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(7)}}</th>
+        <th>{{getMonsterName(7)}}</th>
         <td>{{formateTueur(tueurs.esprits)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(4)}}</th>
+        <th>{{getMonsterName(4)}}</th>
         <td>{{formateTueur(tueurs.humains)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(8)}}</th>
+        <th>{{getMonsterName(8)}}</th>
         <td>{{formateTueur(tueurs.insectes)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(5)}}</th>
+        <th>{{getMonsterName(5)}}</th>
         <td>{{formateTueur(tueurs.machines)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(10)}}</th>
+        <th>{{getMonsterName(10)}}</th>
         <td>{{formateTueur(tueurs.plantes)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(9)}}</th>
+        <th>{{getMonsterName(9)}}</th>
         <td>{{formateTueur(tueurs.pierres)}}</td>
       </tr>
       <tr>
-        <th>{{formateMonsterType(11)}}</th>
+        <th>{{getMonsterName(11)}}</th>
         <td>{{formateTueur(tueurs.mortsVivants)}}</td>
       </tr>
     </table>
