+++ src/app/ffbe/objets-display-components/objet-tueurs-physiques-magiques-display/objet-tueurs-physiques-magiques-display.component.css
