+++ src/app/ffbe/objets-display-components/objet-tueurs-physiques-magiques-display/objet-tueurs-physiques-magiques-display.component.html
@@ -5,18 +5,18 @@
 
   <div *ngIf="displayed">
 
-    <div [ngClass]="'objet-tueurs-left'">
+    <div class="objet-tueurs-left">
       <app-objet-tueurs-display [titre]="'Physiques'"
                                 [tueurs]="objet.tueursPhysiques">
       </app-objet-tueurs-display>
     </div>
 
-    <div [ngClass]="'objet-tueurs-right'">
+    <div class="objet-tueurs-right">
       <app-objet-tueurs-display [titre]="'Magiques'"
                                 [tueurs]="objet.tueursMagiques">
       </app-objet-tueurs-display>
     </div>
-    <div [ngClass]="'objet-tueurs-clear'"></div>
+    <div class="objet-tueurs-clear"></div>
 
   </div>
 </mat-card>
